requirements.txt
Django==3.1.5
djangorestframework==3.12.2
gTTS==2.2.1
pyttsx3==2.90
requests==2.25.1
